refactored code
What changes you have made?
.....
changed names
added stream handling

Why this change is necessary?
....
this was necessary for better understanding of code
stream handling was necessary for handling error and out stream

What is the impact of the change?
.....
this will help to segregate between error stream and out stream